+++ packages/admin/cms-admin/src/dashboard/widgets/DashboardWidgetRoot.tsx
@@ -2,18 +2,11 @@ import { Grid, Paper, Typography } from "@mui/material";
 import { styled } from "@mui/material/styles";
 import * as React from "react";
 
-type Props = {
+export type DashboardWidgetRootProps = React.PropsWithChildren<{
     header: React.ReactNode;
-};
+}>;
 
-const HeaderWrapper = styled("div")`
-    padding: 16px;
-    background-color: ${({ theme }) => theme.palette.grey["A200"]};
-    color: white;
-`;
-
-export const DashboardWidgetContainer: React.FC<React.PropsWithChildren<Props>> = (props: React.PropsWithChildren<Props>) => {
-    const { header, children } = props;
+export const DashboardWidgetRoot = ({ header, children }: DashboardWidgetRootProps) => {
     return (
         <Grid item xs={12} lg={6}>
             <Paper square={false}>
@@ -25,3 +18,9 @@ export const DashboardWidgetContainer: React.FC<React.PropsWithChildren<Props>> 
         </Grid>
     );
 };
+
+const HeaderWrapper = styled("div")`
+    padding: 16px;
+    background-color: ${({ theme }) => theme.palette.grey["A200"]};
+    color: white;
+`;