+++ packages/api/cms-api/src/generator/generate-crud-input.ts
@@ -90,12 +90,7 @@ export async function generateCrudInput(
             type = "Date";
         } else if (prop.type === "BooleanType" || prop.type === "boolean") {
             const initializer = morphTsProperty(prop.name, metadata).getInitializer()?.getText();
-            const defaultValue =
-                prop.nullable && (initializer == "undefined" || initializer == "null")
-                    ? "null"
-                    : initializer == "true" || initializer == "false"
-                    ? initializer
-                    : undefined;
+            const defaultValue = prop.nullable && (initializer == "undefined" || initializer == "null") ? "null" : initializer;
             const fieldOptions = tsCodeRecordToString({ nullable: prop.nullable ? "true" : undefined, defaultValue });
             decorators.push("@IsBoolean()");
             decorators.push(`@Field(${fieldOptions})`);