+++ packages/admin/admin-color-picker/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @comet/admin-color-picker
 
+## 4.4.1
+
+### Patch Changes
+
+-   Updated dependencies [662abcc9]
+    -   @comet/admin@4.4.1
+    -   @comet/admin-icons@4.4.1
+
 ## 4.4.0
 
 ### Minor Changes